Addressing PR feedback; factored out code to add Tools to LoamGraphs; factored out Csv-transformation methods from IntakeSyntax; added AGGREGATOR_INTAKE_DO_UPLOAD sysprop-flag to UkbbDietaryGwas; removed unused schema-file and list-file generation from IntakeSyntax
