--- src/main/scala/loamstream/loam/intake/Schema.scala
@@ -1,7 +1,0 @@
-package loamstream.loam.intake
-
-/**
- * @author clint
- * Jan 3, 2020
- */
-final case class Schema(columns: Seq[(ColumnName, DataType)])
